swap out context for use effect
