+++ packages/tupaia-web/src/features/Map/MapOverlaySelector/MapOverlayDatePicker.tsx
@@ -11,7 +11,6 @@ import { useMapOverlayMapData } from '../utils';
 import { DateRangePicker } from '../../../components';
 import { useDateRanges } from '../../../utils';
 import { URL_SEARCH_PARAMS } from '../../../constants';
-import { useMapOverlaySelectorContext } from './MapOverlaySelectorContext';
 
 const Wrapper = styled.div`
   background-color: ${({ theme }) => theme.overlaySelector.overlayNameBackground};
@@ -23,7 +22,6 @@ const Wrapper = styled.div`
 export const MapOverlayDatePicker = () => {
   const { projectCode, entityCode } = useParams();
   const { selectedOverlay } = useMapOverlays(projectCode, entityCode);
-  const { setMapOverlayDateRange } = useMapOverlaySelectorContext();
   const {
     showDatePicker,
     startDate,
@@ -36,16 +34,6 @@ export const MapOverlayDatePicker = () => {
   } = useDateRanges(URL_SEARCH_PARAMS.MAP_OVERLAY_PERIOD, selectedOverlay);
 
   const { isLoading: isLoadingMapOverlayData } = useMapOverlayMapData();
-
-  const handleSetDate = (_startDate: string, _endDate: string) => {
-    const saveMapOverlayDateRangeForCode = dateRangeString => {
-      setMapOverlayDateRange(selectedOverlay.code, dateRangeString);
-    };
-
-    if (typeof setDates === 'function') {
-      setDates(_startDate, _endDate, saveMapOverlayDateRangeForCode);
-    }
-  };
 
   if (!showDatePicker) return null;
   return (
@@ -62,7 +50,7 @@ export const MapOverlayDatePicker = () => {
           minDate={minStartDate}
           maxDate={maxEndDate}
           granularity={periodGranularity}
-          onSetDates={handleSetDate}
+          onSetDates={setDates}
           onResetDate={onResetDate}
         />
       )}
